mavlink: move PING stream to separate file
